refactor(common/util): move to common
