+++ src/main/java/com/jiuzhang/flashsale/service/impl/OrderServiceImpl.java
@@ -7,6 +7,7 @@ import javax.annotation.Resource;
 import com.alibaba.fastjson.JSON;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.jiuzhang.flashsale.common.enums.OrderStatus;
+import com.jiuzhang.flashsale.common.util.SnowFlake;
 import com.jiuzhang.flashsale.entity.ActivityEntity;
 import com.jiuzhang.flashsale.entity.OrderEntity;
 import com.jiuzhang.flashsale.exception.OrderCreateException;
@@ -17,7 +18,6 @@ import com.jiuzhang.flashsale.exception.OrderPayException;
 import com.jiuzhang.flashsale.mapper.ActivityMapper;
 import com.jiuzhang.flashsale.mapper.OrderMapper;
 import com.jiuzhang.flashsale.service.OrderService;
-import com.jiuzhang.flashsale.util.SnowFlake;
 
 import org.springframework.stereotype.Service;
