+++ src/test/java/com/jiuzhang/flashsale/mapper/OrderMapperTest.java
@@ -8,8 +8,8 @@ import java.time.LocalDateTime;
 import javax.annotation.Resource;
 
 import com.jiuzhang.flashsale.common.enums.OrderStatus;
+import com.jiuzhang.flashsale.common.util.SnowFlake;
 import com.jiuzhang.flashsale.entity.OrderEntity;
-import com.jiuzhang.flashsale.util.SnowFlake;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
